requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pillow>=10.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
rembg>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
fal-client>=0.5.0
python-dotenv>=0.19.0